amirage terminé

+++ src/appendmatrixprovider.h
@@ -14,7 +14,7 @@ class AppendMatrixProvider : public ObservationProvider
 public:
 	AppendMatrixProvider(std::shared_ptr<MatrixLoader> loader, std::shared_ptr<Observation> observation);
 	
-    std::vector< boost::numeric::ublas::matrix< int > > getObservation() const override;
+    std::vector< boost::numeric::ublas::matrix< double > > getObservation() const override;
 	
 private:
 	std::shared_ptr<MatrixLoader> loader;
